Fixed 64 bit pointer issue.
2001-09-20  Akira TAGOH <tagoh@gnome-db.org>

	* sqlite/tool/lemon.c: Fixed 64 bit pointer issue.

+++ providers/gda-default-server/sqlite/tool/lemon.c
@@ -1327,7 +1327,7 @@ char **argv;
 /*
 ** Return a pointer to the next structure in the linked list.
 */
-#define NEXT(A) (*(char**)(((int)A)+offset))
+#define NEXT(A) (*(char**)(((long)A)+offset))
 
 /*
 ** Inputs:
